No commits or merges as i just took existing code from exercise 1 and 2.
Removed any blocking code.

Removed all code from the getFeignData function in FeignController and just returned the JSON object retrieved by feignService.fetchData(). Also changed the methods data type from String to TodoResponse as we were returning a todo object.

Added a for each loop to iterate through the objects in an arraylist that was created to store them.
Once a match was found, I printed the match and then stored that object into a todo object. Which was then returned the talend api.

Converted the typing of TodoClient to an arraylist so i could create an arraylist of TodoResponse type.